[Block Library - Query Loop]: Tidy edit function
